Add Dockerfile_v5.7.1
Note, I based it on `clawpack/v5.7.0_dockerimage:release` because
the `pip install clawpack` kept giving errors when trying to start
from scratch.  Seems to be some problem with gfortran and f2py, maybe in
`jupyter/base-notebook`?

I also merged in the packages needed for GeoClaw so there is only
one version, rather than a separate geoclaw version.

The apps repository is now cloned, rather than downloading a tar file
that seemed to be out of date.